ColorMatchPluginInstaller/aces_color_module.py
# aces_color_module.py
import sys
import subprocess
import os

def install_dependencies():
    """
    Installs the necessary dependencies using pip.
    """
    required_packages = ["OpenColorIO"]
    for package in required_packages:
        subprocess.run([sys.executable, "-m", "pip", "install", package], check=True)
    print("Dependencies installed.")

try:
    import OpenColorIO as ocio  # type: ignore
except ImportError:
    print("OpenColorIO is not installed. Installing now...")
    install_dependencies()
    import OpenColorIO as ocio  # type: ignore

def create_virtual_environment():
    """
    Creates a virtual environment in the current directory.
    """
    subprocess.run([sys.executable, "-m", "venv", ".venv"], check=True)
    print("Virtual environment created.")

def activate_virtual_environment():
    """
    Activates the virtual environment and installs OpenColorIO.
    """
    activate_script = os.path.join(".venv", "Scripts", "activate") if os.name == 'nt' else os.path.join(".venv", "bin", "activate")
    if os.name == 'nt':
        subprocess.run(f"{activate_script} && pip install OpenColorIO", shell=True, check=True)
    else:
        subprocess.run(f"source {activate_script} && pip install OpenColorIO", shell=True, check=True)
    print("Virtual environment activated and OpenColorIO installed.")

def set_ocio_environment_variable(config_path):
    """
    Sets the OCIO environment variable to the given config path.
    """
    os.environ["OCIO"] = config_path
    print(f"OCIO environment variable set to {config_path}")

def apply_aces_color_management(image):
    """
    Applies ACES color management to the given image.

    Args:
        image: The image to which ACES color management will be applied.
    """
    # Create a config object
    config = ocio.Config.CreateFromEnv()

    # Get the processor for the ACES transformation
    processor = config.getProcessor(ocio.ROLE_SCENE_LINEAR, 'ACES - ACEScg')

    # Apply the ACES transformation
    aces_image = processor.applyRGB(image)
    return aces_image

def private_function():
    """
    Private function for internal use.
    """
    print("This is a private function.")

if __name__ == "__main__":
    # Ensure the virtual environment is set up and activated before running this script
    create_virtual_environment()
    activate_virtual_environment()
    set_ocio_environment_variable("C:\\path\\to\\your\\config.ocio")  # Update this path as needed

    example_image = [1.0, 1.0, 1.0]  # Example image data
    result_image = apply_aces_color_management(example_image)
    print(result_image)